fix: resolve all audit findings
Security:
- XSS in GetBucketsJSON: now uses template.HTMLEscapeString + strings.Builder
- Security headers middleware: X-Content-Type-Options, X-Frame-Options, etc.
- Error messages no longer leak internal details to client
- context.Background() replaced with r.Context() in all handlers
- Render errors logged instead of silently discarded

Bugs:
- Encryption audit uses ServerSideEncryption field, not StorageClass
- Access audit false WARN on nil PublicAccessBlock
- Versioning analyzer returns error instead of silent empty result
- Naming analyzer returns partial results on pagination error
- All deep scan errors logged in coordinator.go (not silently discarded)
- DuplicateGroup.Count changed int→int64 for overflow safety

Engine:
- Worker objChan send uses select+ctx.Done (no more blocking forever)
- config/client reads use RLock (no data race)
- Collector panic is recovered (no hanging on <-collCh)
- ListObjectKeys error logged for incremental delta
- SaveScan/AddScanToBucketIndex errors checked and logged
- Dead code removed: retryWithBackoff, RenderString, startTime field

Store:
- BadgerDB value log GC runs every 10 minutes in background
- AddScanToBucketIndex unmarshal error now handled
- DeleteScan cleans bucket index (no orphaned IDs)
- Bucket index keys use separator to prevent prefix collisions

+++ internal/store/models.go
@@ -107,7 +107,7 @@ type DeltaReport struct {
 
 type DuplicateGroup struct {
 	ETag      string   `json:"etag"`
-	Count     int      `json:"count"`
+	Count     int64    `json:"count"`
 	TotalSize int64    `json:"total_size"`
 	Keys      []string `json:"keys"`
 }
